Add @Mappings in interface EventMapper

+++ src/main/java/com/github/java4wro/dto/EventDTO.java
@@ -3,22 +3,23 @@ package com.github.java4wro.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
 public class EventDTO {
 
-    private Integer eventPrice;
+    private BigDecimal eventPrice;
     private String eventName;
     private String eventDescription;
     private LocalDateTime eventDateTime;
 
-    public Integer getEventPrice() {
+    public BigDecimal getEventPrice() {
         return eventPrice;
     }
 
-    public void setEventPrice(Integer eventPrice) {
+    public void setEventPrice(BigDecimal eventPrice) {
         this.eventPrice = eventPrice;
     }
